+++ gulpfile.js
@@ -54,7 +54,7 @@ gulp.task("lint", function (done) {
     "--ignore-pattern",
     "lib/ThirdParty",
     "--max-warnings",
-    "238" // TODO: Bring this back to 0
+    "0"
   ]);
 
   done();