+++ lib/ReactViews/Tour/tour-helpers.ts
@@ -2,10 +2,7 @@
 import { TourPoint } from "../../ReactViewModels/defaultTourPoints";
 import isDefined from "../../Core/isDefined";
 
-import {
-  RelativePosition,
-  TOUR_WIDTH
-} from "../../ReactViewModels/defaultTourPoints";
+import { RelativePosition } from "../../ReactViewModels/defaultTourPoints";
 export {
   RelativePosition,
   TOUR_WIDTH