+++ test/Models/Catalog/CatalogItems/Cesium3DTilesCatalogItemSpec.ts
@@ -340,7 +340,7 @@ describe("Cesium3DTilesCatalogItemSpec", function () {
       "color"
     ]);
     item.setFeatureVisibility(feature, false);
-    // @ts-expect-error
+    // @ts-expect-error JsonValue
     expect(item.style.show.conditions).toEqual([
       ['${color} === "red" && ${doorNumber} === 10', false],
       ["true", true] // fallback rule